Avoid showing a horizontal scrollbar on the home page (#34)
With the body set to `width: 100vw`, browsers with non-overlay
scrollbars showed a horizontal scrollbar on the home page when there
was also vertical scrolling, since `100vw` is the width of the window,
including the vertical scrollbar. Since the body by default has the
width of the window, accounting for scrollbars, removing it will fix
this issue.