Use the Reading info official Lume plugin (#48)
We used to use `reading_time` from the experimental Lume plugins
repository (https://github.com/lumeland/experimental-plugins) to show
an estimate of the reading time needed for a post. While this
experimental plugin still exists in that repo, now it is available as
an official Lume plugin (https://lume.land/plugins/reading_info/), and
therefore we are switching to it.